refactor(commands): extract blue/green + pipeline paths from deploy_run.go
401-line runDeploy split: the terminal, self-contained blue/green canary path
and T05 control-plane pipeline path moved to deploy_run_bluegreen.go (64 lines)
and deploy_run_pipeline.go (133 lines) as handled/err-returning helpers.
deploy_run.go now 269 lines. Supersedes the file's prior CLI-R12 'cannot be
split' note. Pure move, same checks/output/errors, verified via go test.
Part of P6-E2-W1-S1-T3.
